requirements.txt
pefile
pyelftools
pyinstaller
colorama
readchar
pyyaml
